requirements.txt
pytest
selene==2.0.0rc9
allure-pytest
requests
python-dotenv
pydantic
pydantic-settings